+++ R/nass_data.R
@@ -51,72 +51,21 @@ nass_data <- function(source_desc = NULL,
   if (format == "XML") stop("XML not supported yet.")
   
   token <- check_key(token)
-
-  # Check to see if year used a logical operator
-  year  <- trimws(year)
-  punct <- grepl("[[:punct:]]", year)
-  if (length(punct) == 0) punct <- FALSE
-  punct_year <- as.numeric(gsub("[[:punct:]]", "", year))
   
-  args <- list(source_desc = source_desc,
-               sector_desc = sector_desc,
-               group_desc = group_desc,
-               commodity_desc = commodity_desc,
-               short_desc = short_desc,
-               domain_desc = domain_desc,
-               domaincat_desc = domaincat_desc,
-               agg_level_desc = agg_level_desc,
-               statisticcat_desc = statisticcat_desc,
-               state_name = state_name,
-               asd_desc = asd_desc,
-               county_name = county_name,
-               region_desc = region_desc,
-               zip_5 = zip_5,
-               watershed_desc = watershed_desc,
-               freq_desc = freq_desc,
-               reference_period_desc = reference_period_desc,
-               format = format)
+  # Pass the arguments through formatting
+  args <- do.call(args_list, as.list(match.call(expand.dots = FALSE)[-1]))
   
-  # Arguments to upper case
-  args <- lapply(args, function(x) if (!is.null(x)) toupper(x))
   
-  count <- do.call(nass_count, append(args, list(year = year)))
+  count <- do.call(nass_count, args)
   
   if (count > 50000) stop(paste0("Query returns ",
                                  prettyNum(count, big.mark = ","),
                                  " records. The limit is 50,000. Subset the ",
                                  "query to fit within limit. See nass_count()"))
   
-  # Conditional year values
-  if (!punct) {
-    args <- append(args, list(year = year))
-  } else if (punct) {
-    # __LE = <= 
-    # __LT = < 
-    # __GT = > 
-    # __GE = >= 
-    # __LIKE = like 
-    # __NOT_LIKE = not like 
-    # __NE = not equal 
-    if (grepl("^=<|^<=", year) | grepl("=>$|>=$", year)) {
-      args <- append(args, list(year__LE = punct_year))
-    }
-    if ((grepl("^<", year) | grepl(">$", year)) & !grepl("=", year)) {
-      args <- append(args, list(year__LT = punct_year))
-    }
-    if (grepl("^=>|^>=", year) | grepl("=<$|<=$", year)) {
-      args <- append(args, list(year__GE = punct_year))
-    }
-    if ((grepl("^>", year) | grepl("<$", year)) & !grepl("=", year)) {
-      args <- append(args, list(year__GT = punct_year))
-    }
-  }
-  
   base_url <- paste0("http://quickstats.nass.usda.gov/api/api_GET/?key=",
                      token, "&")
-
-  full_url <- httr::modify_url(base_url, query = args)
-  temp     <- httr::GET(full_url)
+  temp     <- httr::GET(base_url, query = args)
   tt       <- check_response(temp)
   
   if (methods::is(tt, "data.frame")) {