+++ NEWS.md
@@ -1,3 +1,7 @@
+# usdarnass 0.1.1
+
+- Arguments can now handle multiple values for an API call.
+
 # usdarnass 0.1.0
 
-* Added a `NEWS.md` file to track changes to the package.
+- Added a `NEWS.md` file to track changes to the package.